requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.44.1
matplotlib==3.8.4
seaborn==0.13.2
nltk==3.8.1
spacy==3.7.4
python-dotenv==1.0.0
joblib==1.3.1
imbalanced-learn==0.11.0
scipy==1.11.4